Fix: include should not have views with wildcard rule based on PR feedback
